pcbnew: Avoid wxassert in message
wxstring::Printf() needs size_t cast to acceptable type or it gives an
assertion in debug mode.

(cherry picked from commit e8ce0ba6ef9de46e2162bb773bc723c77cee40a4)

+++ pcbnew/dialogs/dialog_track_via_properties.cpp
@@ -318,9 +318,9 @@ bool DIALOG_TRACK_VIA_PROPERTIES::confirmPadChange( const std::vector<D_PAD*>& c
     }
     else
     {
-        msg.Printf( _( "This will change the net assigned to %d connected pads to %s.\n"
+        msg.Printf( _( "This will change the net assigned to %lu connected pads to %s.\n"
                        "Do you wish to continue?" ),
-                    changingPads.size(),
+                    static_cast<unsigned long>( changingPads.size() ),
                     m_netSelector->GetValue() );
     }
